#19 Implement Autolayouting config update on cutsom GLSP action
