+++ glsp-server/src/lms/lms-configuration.ts
@@ -1,0 +1,6 @@
+import { injectable } from 'inversify';
+
+@injectable()
+export class LmsConfiguration {
+    public autolayouting = false;
+}
